chore: 并行会话产物收敛 + intent stats 修复 + 注册表同步
并行会话(停止后接手,5min quiesce 确认)遗留产物原样提交:
- R-M03-3 后续 + 研究挖矿模块修改(mining/* 重构)
- adapter TradingRuleChange.field→field_name 重构(测试同步)
- 执行意图/命令聚合等核心文件修改
- pyproject/serde 等配套调整

本会话修复:
- intent.py stats memory 分支 UnboundLocalError(变量名笔误
  state_counts→state_counts_memory,测试锁定)
- 注册表重建(digests/entries/findings 重算,PASS + oracle 0 issues)
- ruff format 统一(5 文件)

门禁实测: 2930 passed / 0 failed、coverage 78.52%、ruff clean、
mypy 0、主扫描器 PASS + 独立 oracle 0 issues

+++ beidou_research/mining/evaluation/multiple_testing.py
@@ -362,7 +362,7 @@ def _spearman_rank_corr(x: list[float], y: list[float]) -> float:
     std_ry = (sum((ry - mean_ry) ** 2 for ry in rank_y) / (n - 1)) ** 0.5
     if std_rx == 0 or std_ry == 0:
         return 0.0
-    return cov / (std_rx * std_ry)
+    return float(cov / (std_rx * std_ry))
 
 
 # ================================================================
